Move Xcode to ios subdirectory
