+++ lib/luffa.rb
@@ -6,4 +6,4 @@ require 'luffa/version'
 require 'luffa/unix_command'
 require 'luffa/with_debugging'
 require 'luffa/retry'
-require 'luffa/xcode'
+require 'luffa/ios/xcode'
