+++ spec/lib/ios/xcode_spec.rb
